seperated results and questions component
