+++ client/src/pages/GetStarted.tsx
@@ -1,6 +1,9 @@
 import Nav from '../components/Navbar';
 import styled from 'styled-components';
 import Questions from '../components/Questions';
+import Results from '../components/Results';
+import { useAppSelector } from '../Hooks/hooks';
+import { RootState } from '../App/store.ts';
 
 const Main = styled.main`
   padding-top: 45px;
@@ -36,17 +39,31 @@ const Main = styled.main`
     color: #404337;
   }
 `
+interface plant {
+  id: string;
+  common_name?: string;
+  default_image?: {
+    thumbnail: string;
+  } | null; 
+  watering?: string;
+}
 
-
-
+function resultsIsReady (results : plant[]){
+  if(results.length > 0){
+    return true;
+  }
+  return false;
+}
 
 const GetStarted = () => {
+  const queryResult : plant[] = useAppSelector((state : RootState) => state.queryResult);
   console.log("render");
   return (
     <div>
       <Nav />
       <Main>
-        <Questions/>
+        {!(resultsIsReady(queryResult)) && <Questions/>}
+        {resultsIsReady(queryResult) && <Results />}
       </Main>
     </div>
   )
